html/semantics/scripting-1/the-script-element/data-url.html

<!doctype html>
<meta charset="utf-8">
<title>data URL and scripts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
  setup({allow_uncaught_exception:true})
  var t = async_test()
  t.step(() => {
    var counter = 0
    window.onerror = t.step_func((message, x, xx, xxx, e) => {
      assert_equals(typeof e, "number")
      assert_not_equals(message, "Script error.")
      if(counter == 2) {
        t.done()
      }
      counter++
    })
  })
</script>
<script src="data:,throw 1"></script>
<script src="data:,throw 2" crossorigin></script>
<script src="data:,throw 3" crossorigin=use-credentials></script>
